+++ packages/adapter-cloudflare-workers/index.js
@@ -16,13 +16,6 @@ import { fileURLToPath } from 'url';
 
 /** @type {import('.').default} */
 export default function ({ config = 'wrangler.toml' } = {}) {
-	// TODO remove for 1.0
-	if (arguments.length > 0) {
-		throw new Error(
-			'esbuild options can no longer be passed to adapter-cloudflare-workers — see https://github.com/sveltejs/kit/pull/4639'
-		);
-	}
-
 	return {
 		name: '@sveltejs/adapter-cloudflare-workers',
 
@@ -63,6 +56,7 @@ export default function ({ config = 'wrangler.toml' } = {}) {
 
 			await esbuild.build({
 				platform: 'browser',
+				conditions: ['worker', 'browser'],
 				sourcemap: 'linked',
 				target: 'es2020',
 				entryPoints: [`${tmp}/entry.js`],