+++ sites/kit.svelte.dev/postcss.config.js
@@ -2,7 +2,7 @@ import { readdirSync } from 'fs';
 import { extname } from 'path';
 import postcssPresetEnv from 'postcss-preset-env';
 
-const baseCSSDir = '../site-kit/src/lib/styles/';
+const baseCSSDir = './node_modules/@sveltejs/site-kit/styles/';
 const cssFiles = readdirSync(baseCSSDir).filter(filename => extname(filename) === '.css').map(filename => baseCSSDir + filename);
 
 export default {