+++ sites/kit.svelte.dev/src/cjs-polyfills.cjs
@@ -1,2 +1,0 @@
-// Polyfill for IntersectionObserver
-require('intersection-observer');