+++ sites/kit.svelte.dev/src/extra-polyfills.js
@@ -1,0 +1,2 @@
+// Polyfill for IntersectionObserver
+import 'intersection-observer';