+++ sites/kit.svelte.dev/src/lib/search/Search.svelte
@@ -9,7 +9,6 @@
 
 <form class="search-container" action="/search">
 	<input
-		id="search"
 		value={q}
 		on:input={(e) => {
 			$searching = true;
@@ -21,14 +20,15 @@
 		type="search"
 		name="q"
 		placeholder="Search{has_search_support ? '' : ' (N/A)'}"
+		aria-label="Search"
 		spellcheck="false"
 		disabled={!has_search_support || undefined}
 	/>
 
 	{#if browser && has_search_support}
-		<label for="#search">
+		<div class="shortcut">
 			<kbd>{navigator.platform === 'MacIntel' ? '⌘' : 'Ctrl'}</kbd> <kbd>K</kbd>
-		</label>
+		</div>
 	{/if}
 </form>
 
@@ -69,7 +69,7 @@
 		background-color: #aaaaaa;
 	}
 
-	input:focus + label {
+	input:focus + .shortcut {
 		display: none;
 	}
 
@@ -78,7 +78,7 @@
 		text-transform: uppercase;
 	}
 
-	label {
+	.shortcut {
 		color: var(--sk-text-3);
 		position: absolute;
 		top: calc(50% - 0.9rem);
@@ -107,7 +107,7 @@
 			width: 11rem;
 		}
 
-		label {
+		.shortcut {
 			padding: 0 1.6rem 0 0;
 		}