+++ sites/kit.svelte.dev/src/lib/search/SearchBox.svelte
@@ -1,6 +1,6 @@
 <script>
 	import { onMount } from 'svelte';
-	import { Icon } from '@sveltejs/site-kit';
+	import Icon from '@sveltejs/site-kit/components/Icon.svelte';
 	import { afterNavigate } from '$app/navigation';
 	import { searching, query, recent } from './stores.js';
 	import { focusable_children, trap } from '../actions/focus.js';