+++ sites/site-kit/src/routes/+page.svelte
@@ -1,1 +1,0 @@
-Hello @sveltejs/site-kit