+++ packages/amp/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sveltejs/amp
 
+## 1.0.0
+
+### Major Changes
+
+First major release, see below for the history of changes that lead up to this.
+Starting from now all releases follow semver and changes will be listed as Major/Minor/Patch
+
+## 1.0.0-next.3
+
+### Patch Changes
+
+- [chore] add peerDependencies, add more specific next version ([#8141](https://github.com/sveltejs/kit/pull/8141))
+
 ## 1.0.0-next.2
 
 ### Patch Changes